requirements.txt
pillow
pafy
tkintertable
urllib3
youtube_dl